arm/arm64: setup: support reset
Supporting reset is surprisingly easy. We just need to zero the
BSS and rerun setup, with the caveat that we shouldn't attempt
to copy the FDT or INITRD from their initial locations again.
We also add a 64K region of memory that will not be cleared on
reset (persistent_memory). Unit tests that plan to reset may
want to make use of that.

Signed-off-by: Andrew Jones <drjones@redhat.com>