scripts/arch-run: don't use env ENV
It turns out that the environment variable 'ENV' is a bad choice
for a name, as it's too basic and can easily conflict with other
uses. Indeed Environment Modules export ENV for their own purposes,
so any users with them enabled will start sending
/usr/share/Modules/init/profile.sh to their unit tests instead of
errata.txt, which will cause setup_env() to spew a bunch of
invalid environment variable warnings, and could cause an unexpected
host crash, depending on what errata the host has and what tests are
being executed. Add a [long] prefix to the variable name to forever
more avoid collisions.

Signed-off-by: Andrew Jones <drjones@redhat.com>