scripts/arch-run: fix bad commit
Commit 171aa3a2 ("Make scripts/arch-run.bash compatible with Bash
4.1 and older") is a commit of the wrong version of the patch.
The correct version to commit was [*]. Fix this up, which restores
the ability to override errata in the errata.txt file with
environment variables.

The erroring out when $commit is not set was changed to just skipping
the line because the erroring out didn't halt the execution of the
test anyway.

[*] https://www.spinics.net/lists/kvm/msg167833.html

Cc: Peter Shier <pshier@google.com>
Signed-off-by: Andrew Jones <drjones@redhat.com>