arm64: Use cpu=max and gic-version=max
When QEMU's cpu type 'max' is available then it's the better choice
for testing, as it implements all supported features. gic-version=max
similarly allows us to test with a more recent GIC. The switch to max
only applies to testing with TCG. For KVM, we still test with 'host'
for each parameter by default. However, for both TCG and KVM we give
the user the chance to override the defaults. To override the cpu
default configure with '--processor' as before (but now it'll apply
to KVM too). To override the gic version set the GIC_VERSION
environment varible, e.g. GIC_VERSION=3. If QEMU doesn't like the
choices, then it'll fail to run, maybe even gracefully.

While touching arm/run, change the 'grep patt >/dev/null's to
'grep -q patt's.

Signed-off-by: Andrew Jones <drjones@redhat.com>